recommendation_api/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
redis>=5.0
faiss-cpu>=1.8
numpy>=1.26
tensorflow>=2.15
pydantic>=2.0
pydantic-settings>=2.0
